ez_wsi_dicomweb/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via -r requirements.in
cachetools==5.5.0
    # via
    #   -r requirements.in
    #   google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
dataclasses-json==0.6.7
    # via -r requirements.in
google-api-core==2.21.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.35.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via -r requirements.in
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.18.2
    # via -r requirements.in
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.65.0
    # via google-api-core
httplib2==0.22.0
    # via google-auth-httplib2
idna==3.10
    # via requests
marshmallow==3.23.0
    # via dataclasses-json
mypy-extensions==1.0.0
    # via typing-inspect
numpy==2.1.2
    # via
    #   -r requirements.in
    #   opencv-python-headless
opencv-python-headless==4.10.0.84
    # via -r requirements.in
packaging==24.1
    # via marshmallow
pillow==11.0.0
    # via -r requirements.in
proto-plus==1.24.0
    # via google-api-core
protobuf==5.28.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==6.1.0
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydicom==3.0.1
    # via -r requirements.in
pyparsing==3.2.0
    # via httplib2
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
    #   requests-mock
    #   requests-toolbelt
requests-mock==1.12.1
    # via -r requirements.in
requests-toolbelt==1.0.0
    # via -r requirements.in
retrying==1.3.4
    # via -r requirements.in
rsa==4.9
    # via google-auth
six==1.16.0
    # via retrying
typing-extensions==4.12.2
    # via typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.3
    # via requests
